+++ plugin.php
@@ -3,7 +3,7 @@
 require_once __DIR__ . "/vendor/autoload.php";
 
 $id = 'hub2';
-$version = '4.5.0';
+$version = '4.6.0';
 $ilias_min_version = '8.12';
 $ilias_max_version = '8.999';
 $responsible = 'sr.solutions';