+++ src/Object/Course/ICourseDTO.php
@@ -1,5 +1,13 @@
 <?php
 
+/*********************************************************************
+ * This Code is licensed under the GPL-3.0 License and is Part of a
+ * ILIAS Plugin developed by sr solutions ag in Switzerland.
+ *
+ * https://sr.solutions
+ *
+ *********************************************************************/
+
 namespace srag\Plugins\Hub2\Object\Course;
 
 use srag\Plugins\Hub2\MappingStrategy\IMappingStrategyAwareDataTransferObject;
@@ -8,6 +16,7 @@ use srag\Plugins\Hub2\Object\DTO\IDidacticTemplateAwareDataTransferObject;
 use srag\Plugins\Hub2\Object\DTO\ITaxonomyAndMetadataAwareDataTransferObject;
 use srag\Plugins\Hub2\Object\DTO\INewsSettingsAwareDataTransferObject;
 use srag\Plugins\Hub2\Object\DTO\ILearningProgressSettingsAwareDataTransferObject;
+use srag\Plugins\Hub2\Object\DTO\ICalendarSettingsAwareDataTransferObject;
 
 /**
  * Interface ICourseDTO
@@ -16,11 +25,12 @@ use srag\Plugins\Hub2\Object\DTO\ILearningProgressSettingsAwareDataTransferObjec
  * @author  Fabian Schmid <fs@studer-raimann.ch>
  */
 interface ICourseDTO extends IDataTransferObject,
-                             ITaxonomyAndMetadataAwareDataTransferObject,
-                             IMappingStrategyAwareDataTransferObject,
-                             IDidacticTemplateAwareDataTransferObject,
-                             INewsSettingsAwareDataTransferObject,
-                             ILearningProgressSettingsAwareDataTransferObject
+    ITaxonomyAndMetadataAwareDataTransferObject,
+    IMappingStrategyAwareDataTransferObject,
+    IDidacticTemplateAwareDataTransferObject,
+    INewsSettingsAwareDataTransferObject,
+    ICalendarSettingsAwareDataTransferObject,
+    ILearningProgressSettingsAwareDataTransferObject
 {
     // @see ilCourseConstants
     public const SUBSCRIPTION_TYPE_DEACTIVATED = 0;